fix: resolve integrated release checks and preserve schedule controls
